lab3/user_sqldb_init/db_init_requirements.txt
passlib==1.7.4
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36